Корректировки координаторов, добавление комментариев
