+++ MdEditor/Presentation/Coordinator/AppCoordinator.swift
@@ -7,35 +7,62 @@
 
 import UIKit
 
-/// Координатор сценариев
-protocol IAppCoordinator: ICoordinator {
-	func showMainFlow()
+// Варианты открытия сценариев
+enum MainFlowOption {
+	case createDocument(createAction: (String) -> Void)
+	case openDirectoryObject
+	case aboutApp
 }
 
-/// Координатор сценариев
+/// Координатор приложения
+protocol IAppCoordinator: ICoordinator {
+	/// Вариант открытия сценариев
+	var mainFlowOption: MainFlowOption? { get set }
+	/// Показывает сценарий стартовой сцены
+	func showStartSceneFlow()
+	/// Показывает сценарий открытия каталогов и файлов
+	func showOpenDocumentFlow()
+}
+
+/// Координатор приложения
 final class AppCoordinator: IAppCoordinator {
 
 	// MARK: Internal properties
 
+	var mainFlowOption: MainFlowOption?
 	weak var finishDelegate: ICoordinatorFinishDelegate?
 	var navigationController: UINavigationController
 	var childCoordinators: [ICoordinator] = []
+
+	// MARK: Lificycle
 
 	init(navigationController: UINavigationController) {
 		self.navigationController = navigationController
 	}
 
-	// MARK: IAppCoordinator
-
-	func showMainFlow() {
-		let mainCoordinator = MainCoordinator(navigationController: navigationController)
-		childCoordinators.append(mainCoordinator)
-		mainCoordinator.start()
-	}
-
 	// MARK: ICoordinator
 
 	func start() {
-		showMainFlow()
+
+		if self.mainFlowOption == nil {
+			showStartSceneFlow()
+		} else {
+			showOpenDocumentFlow()
+		}
+	}
+
+	// MARK: IMainCoordinator
+
+	func showStartSceneFlow() {
+		let startSceneCoordinator = StartSceneCoordinator(navigationController: navigationController)
+		childCoordinators.append(startSceneCoordinator)
+		startSceneCoordinator.start()
+	}
+
+	func showOpenDocumentFlow() {
+		let openDocumentCoordinator = OpenDocumentCoordinator(navigationController: navigationController)
+		childCoordinators.append(openDocumentCoordinator)
+		openDocumentCoordinator.mainFlowType = mainFlowOption
+		openDocumentCoordinator.start()
 	}
 }
